requirements.txt
aiofiles
aiohttp
apscheduler
asyncio
bard
beautifulsoup4
bing_image_downloader
bs4
cloudscraper
curl-cffi
deepai
dnspython
emojis
Faker
ffmpeg-python
flask
future
g4f
geopy
gitpython
google-api-python-client
gpytranslate
gtts
hachoir
heroku3
httpx[http2]==0.27.2
img2pdf
instaloader
langdetect
lexica-api
motor
nekosbest
numpy
openai==0.28.0
opencv-python
pickledb
pillow==10.3.0
psutil
lexica-api
ntgcalls
pytgcalls==3.0.0.dev20
pycountry
pycryptodome
pydub
pyfiglet
pykeyboard
pyshorteners
python-dotenv
python-telegram-bot
python-whois
pyttsx3
pytube
pytz
pyyaml
qrcode
requests
search_engine_parser
speedtest-cli
SpeechRecognition
spotipy
telegraph
tgcrypto
unidecode
wget
youtube-search
youtube-search-python
pyrogram==2.0.106
git+https://github.com/yt-dlp/yt-dlp.git
